requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-cors==0.8.1
aiosignal==1.4.0
alfworld==0.4.2
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.71.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
apache-tvm-ffi==0.1.3
astor==0.8.1
attrs==25.4.0
audioread==3.1.0
av==16.0.1
beautifulsoup4==4.14.3
blake3==1.0.8
blinker==1.9.0
blis==1.3.3
cachetools==6.2.2
catalogue==2.0.10
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.2.1
clip==1.0
cloudpathlib==0.23.0
cloudpickle==3.1.2
codetiming==1.4.0
colorful==0.5.8
compressed-tensors==0.10.2
confection==0.1.5
contourpy==1.3.3
cramjam==2.11.0
cuda-bindings==13.0.3
cuda-pathfinder==1.3.2
cuda-python==13.0.3
cupy-cuda12x==13.6.0
cycler==0.12.1
cymem==2.0.13
datasets==4.4.1
decorator==5.2.1
depyf==0.19.0
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
einops==0.8.1
email-validator==2.3.0
faiss-cpu==1.13.0
Farama-Notifications==0.0.4
fast_downward_textworld==20.6.4
fastapi==0.122.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.2
fastar==0.8.0
fastparquet==2025.12.0
fastrlock==0.8.3
filelock==3.19.1
flash_attn==2.8.3
flashinfer-python==0.5.2
Flask==3.1.3
fonttools==4.61.0
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
gdown==5.2.1
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-auth==2.47.0
google-genai==0.6.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
gym==0.26.2
gym-notices==0.1.0
gym-sokoban==0.0.6
gymnasium==1.2.2
h11==0.16.0
hashids==1.3.1
hf-xet==1.3.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.34.4
hydra-core==1.3.2
identify==2.6.15
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
interegular==0.3.3
itsdangerous==2.2.0
jericho==3.3.1
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lark==1.2.2
latex2sympy2_extended==1.10.2
lazy_loader==0.4
librosa==0.11.0
liger_kernel==0.6.4
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.12
loguru==0.7.3
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
math-verify==0.8.0
matplotlib==3.10.7
mdurl==0.1.2
mementos==1.3.1
mistral_common==1.8.5
model-hosting-container-standards==0.1.9
modelscope==1.33.0
more-itertools==10.8.0
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.0
multiprocess==0.70.18
murmurhash==1.0.15
networkx==3.5
ninja==1.13.0
nodeenv==1.9.1
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cudnn-frontend==1.16.0
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-cutlass-dsl==4.3.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
openai==1.90.0
openai-harmony==0.0.8
openai-whisper==20250625
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
opentelemetry-api==1.38.0
opentelemetry-exporter-prometheus==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
outlines_core==0.2.10
packaging==25.0
pandas==2.3.3
partial-json-parser==0.2.1.1.post7
peft==0.18.0
pillow==11.3.0
pip==25.3
platformdirs==4.5.0
pooch==1.8.2
pre_commit==4.5.0
preshed==3.0.12
proglog==0.1.12
prometheus_client==0.23.1
prometheus-fastapi-instrumentator==7.1.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.1.3
py-cpuinfo==9.0.0
py-spy==0.4.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pybind11==3.0.1
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-extra-types==2.10.6
Pygments==2.19.2
pylatexenc==2.10
PyMuPDF==1.27.2
pyparsing==3.2.5
PySocks==1.7.1
pysubs2==1.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
pyvers==0.1.0
PyYAML==6.0.3
pyzmq==27.1.0
qwen-omni-utils==0.0.8
qwen-vl-utils==0.0.14
ray==2.52.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rpds-py==0.29.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.2.3
sentencepiece==0.2.1
sentry-sdk==2.46.0
setproctitle==1.3.7
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.4
srsly==2.5.2
starlette==0.50.0
supervisor==4.3.0
sympy==1.14.0
tabulate==0.9.0
TatSu==5.8.3
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensordict==0.9.1
termcolor==3.2.0
textworld==1.6.2
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.7.1
torchaudio==2.7.1
torchdata==0.11.0
torchvision==0.22.1
tqdm==4.67.1
transformers==4.57.3
triton==3.3.1
typer==0.20.0
typer-slim==0.20.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
virtualenv==20.35.4
vllm==0.10.0
wandb==0.23.0
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.2.14
weasel==0.4.3
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
xformers==0.0.31
xgrammar==0.1.21
xxhash==3.6.0
yarl==1.22.0
yt-dlp==2025.11.12
zipp==3.23.0

